Add container_put_headers to ssbench
This patch allows you to specify container PUT headers that
will be passed to containers being created.

A new entry to the scenario file `container_put_headers` has
been added, and allows a user to specify any extra headers that
each container will be sent on creation.

This is useful if you want to use ssbench to test something like
container sharding/deep containers as it means you can enable containers
with:

  {                                                                                                                                                                                 "name": "Sharding scenario",
    "sizes": [{
      ...
    }],
    "initial_files": {
      ...
    },
    ...
    "container_base": "shardme",
    "container_count": 1,
    "container_put_headers": {
      "X-Container-Sharding": "on"
    }
  }